+++ NEWS.md
@@ -6,6 +6,9 @@
 
 - Better labelling for `vcov_type` when the user supplies a string.
   “classical” becomes “Classical” instead of “IID”, etc.
+- Improve documentation for `dvnames(skip=)`. Thanks to @raffaem for
+  report \#933.
+- Documentation improvements.
 
 ## 2.5.0